Added averages for each graph - fixed typo in yearly graph
Each graph contains its own average temperature value pertaining to the amount of time the graph tracks (eg: hourly graph shows last hours average, weekly graph shows the last weeks average temperature).
I also fixed a typo in the yearly graph that was causing it to be completely dysfunctional. "--start", "-1y" is proper for starting a year in the past, "-1yh" is not. :thumbsup: