Added SMS support/logging
Added SMS to be sent when temperatures are read to be too high. Also added logging to be able to tell when the last SMS was sent so that we are not spammed with SMS messages when a problem occurs (SMS are set to only send after 1 hour has passed since the last SMS was sent.) This is the case, even if temperatures are still over alert threshold.